Update completeEnrolment.sh
just making sure

+++ completeEnrolment.sh
@@ -1526,7 +1526,9 @@ case $1 in
    TEMP_ADMIN_HOME="$( dscl . read "/Users/$TEMP_ADMIN" NFSHomeDirectory | awk -F ': ' '{ print $NF }' )"
    logIt "Removing $TEMP_ADMIN as no longer required."
    runIt "sysadminctl -deleteUser '$TEMP_ADMIN' -adminUser '$TEMP_ADMIN' -adminPassword '$( readSaved laps )'" "Deleting user '$TEMP_ADMIN'"
-   runIt "rm -rf '$TEMP_ADMIN_HOME'" "Removing '$TEMP_ADMIN' home folder: $TEMP_ADMIN_HOME"
+   if [ -e "$TEMP_ADMIN_HOME" ]; then
+    runIt "rm -rf '$TEMP_ADMIN_HOME'" "Removing '$TEMP_ADMIN' home folder: $TEMP_ADMIN_HOME"
+   fi
   fi
   logIt "Removing: $CLEANUP_FILES"
   eval "rm -rf $CLEANUP_FILES"
